Failure to commit offset should not continue to throw an exception upwards, causing the task to restart (#75)
### Motivation

Failure to commit offset should not continue to throw an exception upwards, causing the task to restart. At present, there will be 3 retry attempts after the commit fails, and the metric of commit-failed counter will be recorded when the number of retry attempts exceeds 3 times. I wonder if we should remove this line of throwing exception. Otherwise, it may be thrown out of the `notifyCheckpointComplete()` function, causing the flink task to fail and restart.

### Expected behaviors

The failure of commiting offsets should not affect the running source task.

### Modifications

Remove one line from `PulsarFetcher`.